work/requirement.txt
matplotlib
numpy
opencv-python
paddleseg
pandas
scipy
tqdm
visdom
